requirements.txt
Django==4.1.7
python-decouple==3.7
requests==2.28.2
gunicorn==20.1.0 
whitenoise==6.4.0
cryptography==39.0.2
APScheduler==3.10.1
jdatetime==4.1.0